// src/src/generated/zod/schemas/objects/InvoiceMaxOrderByAggregateInput.schema.ts
import { z } from 'zod';
import { SortOrderSchema } from '../enums/SortOrder.schema';

import type { Prisma } from '@prisma/client';

const Schema: z.ZodType<Prisma.InvoiceMaxOrderByAggregateInput> = z
  .object({
    id: z.lazy(() => SortOrderSchema).optional(),
    total: z.lazy(() => SortOrderSchema).optional(),
    issuedDate: z.lazy(() => SortOrderSchema).optional(),
    dueDate: z.lazy(() => SortOrderSchema).optional(),
    status: z.lazy(() => SortOrderSchema).optional(),
    createdAt: z.lazy(() => SortOrderSchema).optional(),
    updatedAt: z.lazy(() => SortOrderSchema).optional(),
    createdById: z.lazy(() => SortOrderSchema).optional(),
    paymentUrl: z.lazy(() => SortOrderSchema).optional(),
    taxPercentage: z.lazy(() => SortOrderSchema).optional(),
    discountPercentage: z.lazy(() => SortOrderSchema).optional(),
  })
  .strict();

export const InvoiceMaxOrderByAggregateInputObjectSchema = Schema;